+++ fpscompress-template-1.21.11/src/main/java/com/mukulramesh/fpscompress/gui/PreFabStatusMenu.java
@@ -98,7 +98,8 @@ public class PreFabStatusMenu extends AbstractContainerMenu {
                             prefabEntity.getFocusedResourceId(),
                             prefabEntity.getAutoNormalizedTicks(),
                             prefabEntity.getUseAutoNormalize(),
-                            prefabEntity.getAutoNormalizedDisplayMode()
+                            prefabEntity.getAutoNormalizedDisplayMode(),
+                            prefabEntity.getPrefabName()
                         )
                     );
                 }